HTN Planning and Task Decomposition (WIP)
Added more comments and regions to the scripts.
Defined a method to to order a collection of tasks by the most optimal to perform.
The HTN class now orders its available tasks by most the optimal to execute; selects the best task to achieve its goal.
Removed the executable task array and related method in the HTN class.