HTN and Task Fixes
Converted some more pseudo code.
Created an empty method to make the AI look around.
A plan is now assigned when a HTN decomposes a task.
Fixed wrong variable being used on line 164 of the HTN class.
Fixed wrong value being returned on lines 84 and 88 of Task.cs.
Fixed a task not being assign on line 22 of PrimitiveTask.cs.