Defined the Follow Action; Working HTN
Defined a basic follow action for the AI.
Created a variant of the standard asset player prefab.
Reversed the order of a HTN's plan before it's used, to ensure tasks are executed in the correct order.
Task postconditions are set to true when it succeeds in executing.
Fixed the HTN not including it's goal as part of its plan.