Follow & Stay Orders; HTN Fixes
Created a UI prefab and some buttons for giving orders.
Setup two orders for the player to give to the AI agent.
Task now have an associated game-object to get data from, based on what the task does.
Fixed two issues where the HTN's current task was not being erased when the plan is cleared, on lines 75-76 of the HTN class.