Player Can Ping Objects; AI Derives Orders from Pings
Added the keybind for pinging to the control scheme asset.
Created a prefab for a world space ping UI element.
Changed the ping's text type to the text mesh pro text type.
Changed the AI agents tasks variables into properties.
Created new orders that the player can give the AI through pinging.
The player can now ping elements in the environment to give the AI orders (WIP).
Created a new constructor for the Condition struct.
Changed HTN planning so no plan is created if all preconditions of the goal are already true.